realestate_dabang/requirements.txt
selenium>=4.20.0
undetected-chromedriver>=3.5.5
beautifulsoup4>=4.12.3
lxml>=5.2.2
pandas>=2.2.2
openpyxl>=3.1.5
loguru>=0.7.2
pydantic>=2.8.2
customtkinter>=5.2.2
python-dotenv>=1.0.1

